+++ quickstarts/maven/spring-boot-watch/src/main/java/org/eclipse/jkube/maven/sample/spring/boot/HelloController.java
@@ -13,18 +13,15 @@
  */
 package org.eclipse.jkube.maven.sample.spring.boot;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-/**
- * @author roland
- */
+@RestController
+public class HelloController {
 
-@SpringBootApplication
-public class Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
+    @RequestMapping("/")
+    public String index() {
+        return "Greetings from Spring Boot";
     }
 
 }